test: split integration test into each command
